Support mithril as backend (#150)
Add mithril spot-bid as backend. Reservation isn't supported because the
job duration cannot be determined in Velda, but may be achieved by
setting a high limit price.

Support pooling: pause instead of terminating to speed up subsequent
launches.

Support RESTARTING state in agent management: Mithril has a long delay
when a spot-bid is paused to actual instance termination, the connection
will be still live for 2-3 mins. If restarted, it has a high chance to
reuse old name in pooling. NEEDS_RESTART state is a combo of DELETING +
PENDING, expect the workload to disconnect first, then reconnect later.

Current start-up time is about 3 mins for cold start, and ~100s for
resuming.